Function gmsh_sys::gmshModelMeshGetNodes

pub unsafe extern "C" fn gmshModelMeshGetNodes(
    nodeTags: *mut *mut usize, 
    nodeTags_n: *mut usize, 
    coord: *mut *mut f64, 
    coord_n: *mut usize, 
    parametricCoord: *mut *mut f64, 
    parametricCoord_n: *mut usize, 
    dim: c_int, 
    tag: c_int, 
    includeBoundary: c_int, 
    returnParametricCoord: c_int, 
    ierr: *mut c_int
)

Get the nodes classified on the entity of dimension dim and tag tag. If tag < 0, get the nodes for all entities of dimension dim. If dim and tag are negative, get all the nodes in the mesh. nodeTags contains the node tags (their unique, strictly positive identification numbers). coord is a vector of length 3 times the length of nodeTags that contains the x, y, z coordinates of the nodes, concatenated: [n1x, n1y, n1z, n2x, ...]. If dim >= 0 and returnParamtricCoord is set, parametricCoord contains the parametric coordinates ([u1, u2, ...] or [u1, v1, u2, ...]) of the nodes, if available. The length of parametricCoord can be 0 or dim times the length of nodeTags. If includeBoundary is set, also return the nodes classified on the boundary of the entity (which will be reparametrized on the entity if dim >= 0 in order to compute their parametric coordinates).