Function gmsh_sys::gmshViewAddModelData

pub unsafe extern "C" fn gmshViewAddModelData(
    tag: c_int, 
    step: c_int, 
    modelName: *const c_char, 
    dataType: *const c_char, 
    tags: *mut usize, 
    tags_n: usize, 
    data: *mut *const f64, 
    data_n: *const usize, 
    data_nn: usize, 
    time: f64, 
    numComponents: c_int, 
    partition: c_int, 
    ierr: *mut c_int
)

Add model-based post-processing data to the view with tag tag. modelName identifies the model the data is attached to. dataType specifies the type of data, currently either "NodeData", "ElementData" or "ElementNodeData". step specifies the identifier (>= 0) of the data in a sequence. tags gives the tags of the nodes or elements in the mesh to which the data is associated. data is a vector of the same length as tags: each entry is the vector of double precision numbers representing the data associated with the corresponding tag. The optional time argument associate a time value with the data. numComponents gives the number of data components (1 for scalar data, 3 for vector data, etc.) per entity; if negative, it is automatically inferred (when possible) from the input data. partition allows to specify data in several sub-sets.