Function gmsh_sys::gmshViewProbe

pub unsafe extern "C" fn gmshViewProbe(
    tag: c_int, 
    x: f64, 
    y: f64, 
    z: f64, 
    value: *mut *mut f64, 
    value_n: *mut usize, 
    step: c_int, 
    numComp: c_int, 
    gradient: c_int, 
    tolerance: f64, 
    xElemCoord: *mut f64, 
    xElemCoord_n: usize, 
    yElemCoord: *mut f64, 
    yElemCoord_n: usize, 
    zElemCoord: *mut f64, 
    zElemCoord_n: usize, 
    ierr: *mut c_int
)

Probe the view tag for its value at point (x, y, z). Return only the value at step step is step is positive. Return only values with numComp if numComp is positive. Return the gradient of the value if gradient is set. Probes with a geometrical tolerance (in the reference unit cube) of tolerance if tolerance is not zero. Return the result from the element described by its coordinates if xElementCoord, yElementCoord and zElementCoord are provided.